+++ src/main/java/com/likelion/backendplus4/talkpick/backend/auth/presentation/controller/AuthController.java
@@ -1,20 +1,18 @@
 package com.likelion.backendplus4.talkpick.backend.auth.presentation.controller;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.likelion.backendplus4.talkpick.backend.auth.application.port.in.AuthServiceUseCase;
 import com.likelion.backendplus4.talkpick.backend.auth.domain.model.AuthUser;
 import com.likelion.backendplus4.talkpick.backend.auth.infrastructure.support.mapper.AuthUserMapper;
-import com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req.CheckAccountDto;
-import com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req.CheckEmailDto;
-import com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req.CheckNicknameDto;
-import com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req.ConfirmCodeDto;
+import com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req.check.CheckAccountDto;
+import com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req.check.CheckEmailDto;
+import com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req.check.CheckNicknameDto;
+import com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req.confirm.ConfirmCodeDto;
 import com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req.RefreshReqDto;
 import com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req.SignInDto;
 import com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req.SignUpDto;
