--- src/main/java/com/likelion/backendplus4/talkpick/backend/auth/presentation/dto/req/CheckAccountDto.java
@@ -1,6 +1,0 @@
-package com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req;
-
-public record CheckAccountDto(
-	String account
-) {
-}
