--- src/main/java/com/likelion/backendplus4/talkpick/backend/auth/presentation/dto/req/ConfirmCodeDto.java
@@ -1,7 +1,0 @@
-package com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req;
-
-public record ConfirmCodeDto(
-	String email,
-	String code
-) {
-}
