+++ src/main/java/com/likelion/backendplus4/talkpick/backend/auth/presentation/dto/req/check/CheckEmailDto.java
@@ -1,0 +1,6 @@
+package com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req.check;
+
+public record CheckEmailDto(
+	String email
+) {
+}
