+++ src/main/java/com/likelion/backendplus4/talkpick/backend/auth/presentation/dto/req/check/CheckNicknameDto.java
@@ -1,0 +1,6 @@
+package com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req.check;
+
+public record CheckNicknameDto (
+	String nickName
+){
+}
