+++ src/main/java/com/likelion/backendplus4/talkpick/backend/auth/presentation/dto/req/confirm/ConfirmCodeDto.java
@@ -1,0 +1,7 @@
+package com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req.confirm;
+
+public record ConfirmCodeDto(
+	String email,
+	String code
+) {
+}
