+++ src/main/java/com/likelion/backendplus4/talkpick/backend/auth/presentation/support/validator/CheckAccountValidator.java
@@ -5,7 +5,7 @@ import org.springframework.util.StringUtils;
 import org.springframework.validation.Errors;
 import org.springframework.validation.Validator;
 
-import com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req.CheckAccountDto;
+import com.likelion.backendplus4.talkpick.backend.auth.presentation.dto.req.check.CheckAccountDto;
 import com.likelion.backendplus4.talkpick.backend.auth.presentation.support.validator.enums.AuthValidationError;
 
 @Component
